requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-heroku==0.3.1
gunicorn==20.0.4
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
typing-extensions==3.7.4.3
zipp==3.4.0
